% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sam.R
\name{get_sam}
\alias{get_sam}
\title{Get a sam directory from stockassessment.org}
\usage{
get_sam(directory, user = "user3")
}
\arguments{
\item{directory}{Name of the directory (assessment)}

\item{user}{Name of the user}
}
\description{
The function copies the whole directory of an assessment run from
stockassessment.org to a local directory
}
\note{
This functions most likely works only in Linux
}

